requirements.txt
Flask
requests
beautifulsoup4
pandas
tldextract
google-generativeai
python-dotenv
